Use auth endpoint to fix failing tests
