+++ stream-video-android-core/src/test/kotlin/io/getstream/video/android/core/ClientAndAuthTest.kt
@@ -41,7 +41,7 @@ class ClientAndAuthTest : TestBase() {
     fun regularUser() = runTest {
         val builder = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
             user = testData.users["thierry"]!!,
             token = testData.tokens["thierry"]!!,
@@ -55,7 +55,7 @@ class ClientAndAuthTest : TestBase() {
     fun anonymousUser() = runTest {
         val builder = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
             user = User(
                 type = UserType.Anonymous,
@@ -74,7 +74,7 @@ class ClientAndAuthTest : TestBase() {
         // API call is getGuestUser or something like that
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
             user = User(
                 id = "guest",
@@ -90,7 +90,7 @@ class ClientAndAuthTest : TestBase() {
     fun subscribeToAllEvents() = runTest {
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
             user = User(
                 id = "guest",
@@ -108,7 +108,7 @@ class ClientAndAuthTest : TestBase() {
     fun subscribeToSpecificEvents() = runTest {
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
             user = User(
                 id = "guest",
@@ -127,10 +127,10 @@ class ClientAndAuthTest : TestBase() {
     fun waitForWSConnection() = runTest {
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.tokens["thierry"]!!,
+            user = testData.users["thierry"]!!,
+            token = authData!!.token,
         ).build()
         assertThat(client.state.connection.value).isEqualTo(ConnectionState.PreConnect)
         val clientImpl = client as StreamVideoImpl
@@ -149,8 +149,8 @@ class ClientAndAuthTest : TestBase() {
             context = context,
             apiKey = "notvalid",
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.tokens["thierry"]!!,
+            user = testData.users["thierry"]!!,
+            token = authData!!.token,
         ).build()
     }
 
@@ -159,10 +159,10 @@ class ClientAndAuthTest : TestBase() {
     fun `test an expired token, no provider set`() = runTest {
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.expiredToken,
+            user = testData.users["thierry"]!!,
+            token = authData!!.token,
         ).build()
 
         val result = client.call("default", "123").create()
@@ -179,10 +179,10 @@ class ClientAndAuthTest : TestBase() {
         StreamVideo.removeClient()
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.expiredToken,
+            user = testData.users["thierry"]!!,
+            token = testData.expiredToken,
             tokenProvider = { error ->
                 testData.tokens["thierry"]!!
             },
@@ -211,10 +211,10 @@ class ClientAndAuthTest : TestBase() {
     fun `two clients is not allowed`() = runTest {
         val builder = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.tokens["thierry"]!!,
+            user = testData.users["thierry"]!!,
+            token = authData!!.token,
         )
         val client = builder.build()
         val client2 = builder.build()
@@ -226,10 +226,10 @@ class ClientAndAuthTest : TestBase() {
         // often you'll want to run the connection task in the background and not wait for it
         val client = StreamVideoBuilder(
             context = context,
-            apiKey = apiKey,
+            apiKey = authData!!.apiKey,
             geo = GEO.GlobalEdgeNetwork,
-            testData.users["thierry"]!!,
-            testData.tokens["thierry"]!!,
+            user = testData.users["thierry"]!!,
+            token = authData!!.token,
         ).build()
         val clientImpl = client as StreamVideoImpl
         client.subscribe {
